moved existing logging to toollog app
removed syslog loglevel config and added tcp support
