+++ src/toollog/urls.py
@@ -1,0 +1,2 @@
+app_name = 'toollog'
+urlpatterns = []
